Fixed weaving of multi-line read-time conditionals.
We now emit read-time conditional forms one line at a time, like
we do with literal strings. To ensure consistency, we've also
removed the (trivial) optimization of eliding the alignment tabs
for simple forms. (Weaving a top-level multi-line string was also
broken due to this optimization, but we never noticed.)