Added COMPILE-TESTS-FILE argument to TANGLE-FILE.
Some tests are easier to write if we don't have to compile them. This
patch adds a keyword for controlling that.

Additionally, moved the two COMPILE-FILE calls inside a
WITH-COMPILATION-UNIT.  Strictly speaking, the compile-time side effects
of compiling the primary tangled file might not persist after the first
COMPILE-FILE, and so the tests file might not be compilable until after
loading the tangled file.  WITH-COMPILATION-UNIT doesn't standardly do
much, but it's meant to be the place where implementations hang any
functionality for compiling multiple things in a single compilation
environment, so in principle that's the right thing to use.  (Of course,
most implementations' file compilers do just munge a global compilation
environment that outlives the compilation job, so this change is mostly
only aesthetic.)